+++ lib/src/cli/commands/check_unused_l10n.dart
@@ -25,12 +25,6 @@ class CheckUnusedL10nCommand extends BaseCommand {
   }
 
   @override
-  void validateCommand() {
-    validateRootFolderExist();
-    validateTargetDirectories();
-  }
-
-  @override
   Future<void> runCommand() async {
     final rootFolder = argResults[FlagNames.rootFolder] as String;
     final classPattern = argResults[FlagNames.l10nClassPattern] as String;
@@ -48,6 +42,7 @@ class CheckUnusedL10nCommand extends BaseCommand {
       folders,
       rootFolder,
       config,
+      sdkPath: findSdkPath(),
     );
 
     return _analyzer
@@ -61,8 +56,7 @@ class CheckUnusedL10nCommand extends BaseCommand {
   void _addFlags() {
     _usesL10nClassPatternOption();
     _usesReporterOption();
-    usesRootFolderOption();
-    usesExcludeOption();
+    addCommonFlags();
   }
 
   void _usesReporterOption() {